ADD BASIC FUNCTIONALITY TO THE APP
I implemented a list for the to-dos of the day using recyclerView.
I implemented data persistence using the ROOM library.
I enabled creating, updating, and adding of tasks to the database.
All tasks can be deleted using the delete icon at the top of the Home screen.
Used liveData to be notified of changes.
In passing arguments between fragments, I used safeArgs from the Navigation component.